### 2019-01-22 #1197 @whedon
[PRE REVIEW]: eddy: Extracting Protoplanetary Disk Dynamics with Python
**Submitting author:** @richteague (<a href="http://orcid.org/0000-0003-1534-5186">Richard Teague</a>)
**Repository:** <a href="https://github.com/richteague/eddy" target ="_blank">https://github.com/richteague/eddy</a>
**Version:** v1.1
