### 2019-01-22 #1197 @arfon
:wave: @richteague - many thanks for your submission to JOSS. Next up we need to find a reviewer or two for this submission. Could you take a look at [this list of potential reviewers](https://bit.ly/joss-reviewers) and suggest a few here that might be suitable? Alternatively, you can suggest names of people not on this list that I can contact.
