### 2019-01-23 #1197 @adam-m-jcbs
:wave: @arfon, @richteague - I am willing and feel competent to review the submission.  However, I am currently quite subscribed through February.  I cannot guarantee I'll be able to make time in my schedule to complete the review until my plate is clearer in March.  I've agreed to do a review I didn't have room for in the past and am trying to not repeat that mistake.  Were I to be selected as the reviewer, I wouldn't be able to promise completion of the review until March 15.  I will attempt to complete it sooner, but am not comfortable promising to have it done sooner than that.

If this constraint is something y'all are both comfortable with and would still like me to do the review, please let me know.  Looks like an interesting project I'd like to learn more about!
