### 2019-01-24 e2e96dc
oversampling of the line profiles
