### 2019-01-25 #1197 @richteague
@arfon - What's the best practice when it comes to further development of the code? I wanted to submit a JOSS paper for it already as there's been quite a bit of work based on the software so far (published and in prep.), but there's definitely some additional features which I'd be interested in implementing. Would I have to wait for the review (i.e. after 15th March) to begin on this, or is it possible to review this version?
