### 2019-01-30 #1220 @arfon
@kevin-flaherty - thanks again for agreeing to review for JOSS.

Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
