### 2019-02-20 #1220 @kevin-flaherty
@arfon - I want to acknowledge that Rich Teague and I did co-author a paper that was published last year (http://adsabs.harvard.edu/abs/2018ApJ...856..117F). This is the only time that we have worked together, and we have not worked together since then, nor do we have any immediate plans to work together on any future projects. While I don't believe that this collaboration will bias my review, I wanted to disclose this collaboration given the COI rules.
