### 2019-02-22 2451c1d
more explict description of the required arrays
