### 2019-02-26 #1220 @kevin-flaherty
@afron I have completed my review, and I recommend eddy for acceptance by JOSS.
